Refactor Bartlett to use monad transformers
What a ride! I'm not yet convinced that using Monad Transformers is a
net win, but I'm willing to give it a little bit more time. The initial
impetus for using transformers is to combine various monadic contexts in
order to simply each Action module. At the time of this writing I think
the opposite is true: the code is moderately more complex.

I'd like to at least add state and logging facilities before rolling
back these changes, which unfortunately means that new features will
take a bit longer to trickle down into mainline. I'll be sure to explain
this to interested parties.