Fix bug where crumb is ignored if not enabled on Jenkins
This would happen in cases where CSRF protection was not enabled but
Bartlett was still trying to use those values in the request header
(causing Jenkins to freak out).

Bartlett now properly checks for a valid-ish header before including it
in the request (although the implementation is still in need of some
serious cleanup).