Banish lazy ByteStrings where possible
The code interacting with ByteString was getting a little hairy, both
int terms of imports and in terms of working with its values. Where
possible these strings have been converted to their strict cousins,
which has the benefit of improving performance.

Eventually I'd like to minimize the use ByteString where possible and
switch to using Text and it's much nicer API.