Refactor Network module in a pretty major way
Basically forcing Either return types on almost everything in the
Bartlett Monad. This allows us to return better errors that can then be
translated into better error messages for the user.

At this point everything is still WIP, there's still quite a lot of work
left to do in order to clean up the implementation. Once that's done
it'll probably be time to cut 2.0 release since this'll standardize how
errors are reported to the user (and therefore break anyone that
happened to be depending on such behavior).

This is a good opportunity to add some better test coverage too, which
will probably involve tackling how to mock/replace wreq requests.

All that is to say that those hoping for a new release of Bartlett will
likely be waiting a while. If you're reading this, feel free to reach
out and see where you can pitch in!