Refactor Bartlett to rely on Reader more
This is a fairly major re-write of Bartlett. The impetus for this change
is to rely on Reader more to simplify type signatures. It's also become
handy for passing in different versions of functions (like execRequest)
which finally opens up the possibilty of testing the actions based on
dummy responses instead of having to setup a local server to test.

The test suite that currently exists is basically defunct, so that'll
need to get updated before a release is cut. So far, though, things
appear to be functioning correctly.